backend/requirements.txt
pydicom
Pillow
numpy
mysql-connector-python
python-dotenv
pylibjpeg
pylibjpeg-libjpeg